chore: rename files to be consistently lowercase
